Use jiffies as a timesync platform source
It might be possible to use hrtimer in the future, so this should be considered temporary. Let's not pull in architecture specific timers though.

In the process of investigating this one, tried to load all of the greybus kernel modules in to my desktop machine (something that should be an obvious starting point) and it crashes the kernel:

[   98.399919] Unable to find a compatible ARMv7 timer
[   98.399928] divide error: 0000 [#1] SMP PTI
[   98.399931] Modules linked in: greybus(OE+) ipt_MASQUERADE nf_nat_masquerade_ipv4 nf_conntrack_netlink nfnetlink xfrm_user xfrm_algo iptable_nat nf_conntrack_ipv4 nf_defrag_ipv4 nf_nat_ipv4 xt_addrtype iptable_filter xt_conntrack nf_nat nf_conntrack libcrc32c br_netfilter bridge stp llc rfcomm aufs overlay pci_stub vboxpci(OE) vboxnetadp(OE) vboxnetflt(OE) vboxdrv(OE) bnep binfmt_misc nls_iso8859_1 joydev input_leds snd_hda_codec_cirrus snd_hda_codec_generic snd_hda_intel intel_rapl snd_hda_codec applesmc x86_pkg_temp_thermal snd_hda_core intel_powerclamp input_polldev snd_hwdep snd_pcm coretemp wl(POE) btusb btrtl btbcm snd_seq_midi btintel bluetooth uvcvideo videobuf2_vmalloc videobuf2_memops snd_seq_midi_event kvm_intel ecdh_generic videobuf2_v4l2 snd_rawmidi videobuf2_core kvm hid_generic videodev
[   98.399966]  bcm5974 hid_apple snd_seq irqbypass media crct10dif_pclmul snd_seq_device crc32_pclmul i915 cfg80211 snd_timer ghash_clmulni_intel drm_kms_helper pcbc snd drm sbs i2c_algo_bit apple_gmux aesni_intel fb_sys_fops acpi_als syscopyarea kfifo_buf sysfillrect aes_x86_64 mei_me sysimgblt crypto_simd mac_hid industrialio glue_helper soundcore mei cryptd sbshc video lpc_ich shpchp intel_cstate intel_rapl_perf apple_bl sch_fq_codel parport_pc ppdev lp parport ip_tables x_tables autofs4 hid_appleir usbhid hid tg3 firewire_ohci ahci ptp firewire_core sdhci_pci libahci pps_core crc_itu_t sdhci thunderbolt
[   98.400005] CPU: 0 PID: 3636 Comm: insmod Tainted: P           OE    4.15.0-52-generic #56-Ubuntu
[   98.400007] Hardware name: Apple Inc. MacBookPro8,2/Mac-94245A3940C91C80, BIOS    MBP81.88Z.0050.B00.1804101331 04/10/18
[   98.400018] RIP: 0010:gb_timesync_init+0x36/0x77 [greybus]
[   98.400019] RSP: 0018:ffffb49e83287c50 EFLAGS: 00010246
[   98.400022] RAX: 00038d7ea4c68000 RBX: 0000000000000000 RCX: 0000000000000000
[   98.400023] RDX: 0000000000000000 RSI: 0000000000000000 RDI: ffff90d41531faf0
[   98.400025] RBP: ffffb49e83287c58 R08: 0000000000000001 R09: 0000000000000477
[   98.400026] R10: ffff90d3efadd400 R11: 0000000000000000 R12: ffffffffc0a2171c
[   98.400028] R13: ffff90d3efadd140 R14: 0000000000000001 R15: ffff90d3e7eacd20
[   98.400030] FS:  00007f3a90fc2540(0000) GS:ffff90d42fa00000(0000) knlGS:0000000000000000
[   98.400032] CS:  0010 DS: 0000 ES: 0000 CR0: 0000000080050033
[   98.400033] CR2: 00007f78e315c088 CR3: 00000003ef8a2001 CR4: 00000000000606f0
[   98.400035] Call Trace:
[   98.400043]  gb_init+0x7e/0xb0 [greybus]
[   98.400048]  do_one_initcall+0x52/0x19f
[   98.400052]  ? __vunmap+0x8e/0xc0
[   98.400055]  ? _cond_resched+0x19/0x40
[   98.400059]  ? kmem_cache_alloc_trace+0x14e/0x1b0
[   98.400062]  ? do_init_module+0x27/0x209
[   98.400067]  do_init_module+0x5f/0x209
[   98.400071]  load_module+0x191e/0x1f10
[   98.400076]  ? ima_post_read_file+0x96/0xa0
[   98.400079]  SYSC_finit_module+0xfc/0x120
[   98.400082]  ? SYSC_finit_module+0xfc/0x120
[   98.400085]  SyS_finit_module+0xe/0x10
[   98.400087]  do_syscall_64+0x73/0x130
[   98.400091]  entry_SYSCALL_64_after_hwframe+0x3d/0xa2
[   98.400093] RIP: 0033:0x7f3a90ae9839
[   98.400094] RSP: 002b:00007ffdc38c0df8 EFLAGS: 00000246 ORIG_RAX: 0000000000000139
[   98.400096] RAX: ffffffffffffffda RBX: 000055be899b0780 RCX: 00007f3a90ae9839
[   98.400098] RDX: 0000000000000000 RSI: 000055be882e8d2e RDI: 0000000000000003
[   98.400099] RBP: 000055be882e8d2e R08: 0000000000000000 R09: 00007f3a90dbc000
[   98.400101] R10: 0000000000000003 R11: 0000000000000246 R12: 0000000000000000
[   98.400102] R13: 000055be899b2e30 R14: 0000000000000000 R15: 0000000000000000
[   98.400104] Code: 68 00 00 00 85 c0 89 c3 74 0e 48 c7 c7 38 40 c2 c0 e8 81 e2 2c d2 eb 4f e8 7a e8 1f 00 31 d2 89 c6 48 b8 00 80 c6 a4 7e 8d 03 00 <48> f7 f6 31 d2 48 c7 c7 58 40 c2 c0 48 89 35 31 9d 20 00 48 89
[   98.400138] RIP: gb_timesync_init+0x36/0x77 [greybus] RSP: ffffb49e83287c50
[   98.400141] ---[ end trace 4ee254b88e5bc62d ]---

The actual divide by zero is occuring in timesync.c::gb_timesync_init() but the root cause is because of gb_timesync_platform_get_clock_rate() returning zero. If it fails to load an ARMv7 architecture-specific timer, then it falls back to cpufreq_get(0), which has (in this case) returned 0.

It's clear that the original author of timesync.c wanted to have a very fine-grained (femtosecond-accurate) measure of time potentially between several interconnected devices.

Let's take a step back. Clock synchronization between two computers without a common reference clock is a hard problem to solve even only down to the microsecond scale. That's why most bench equipment has an input for the 10 MHz GPS reference clock, has a TXCO, and then contains further IP to synchronize to the GPS clock.

Let's look at one industry where that is particularly true: Software Defined Radio. I've designed picosecond-accurate time synchronisation system between two computers, and I'll say this: unless we've got a hardware doing the synchronisation for us, a Gbps-capable interconnect with extremely predictable latency, with a control loop in place and an extremely accurate reference clock on one end, synchronising clocks of even two different systems will be unreliable.

In the degenerate case where the interconnect is multi-user wireless technology, millisecond-accurate time synchronisation can be obtained (using e.g. NTP).

On wired ethernet systems with IEEE 1588-2002 hardware for Precision Time Protocol (PTP), sub-microsecond accuracy can be obtained for the purposes of running control loops and such.

With that in mind, for the purposes of a generic greybus, it's probably best to use do_gettimeofday() as a baseline, or even jiffies.